+++ demos/frontend/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"encoding/json"
 	_ "github.com/lib/pq"
 	"github.com/pirsch-analytics/pirsch/v4/db"
 	"github.com/pirsch-analytics/pirsch/v4/tracker"
@@ -13,6 +14,7 @@ import (
 // For more details, take a look at the backend demo and documentation.
 func main() {
 	copyPirschJs()
+	copyPirschEventsJs()
 
 	// Set the key for SipHash.
 	tracker.SetFingerprintKeys(42, 123)
@@ -49,10 +51,34 @@ func main() {
 		log.Println("Counted one hit")
 	}))
 
+	// Create an endpoint to handle client event requests.
+	http.Handle("/event", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		var req struct {
+			Name     string            `json:"event_name"`
+			Duration uint32            `json:"event_duration"`
+			Meta     map[string]string `json:"event_meta"`
+		}
+		decoder := json.NewDecoder(r.Body)
+
+		if err := decoder.Decode(&req); err != nil {
+			log.Printf("Error decoding event request: %s", err)
+			return
+		}
+
+		data := tracker.EventOptions{
+			Name:     req.Name,
+			Duration: req.Duration,
+			Meta:     req.Meta,
+		}
+
+		pirschTracker.Event(r, data, tracker.HitOptionsFromRequest(r))
+		log.Println("Received event")
+	}))
+
 	// Add a handler to serve index.html and pirsch.js.
 	http.Handle("/", http.FileServer(http.Dir("./")))
 
-	log.Println("Starting server on port 8080...")
+	log.Println("Started server on http://localhost:8080")
 	log.Fatal(http.ListenAndServe(":8080", nil))
 }
 
@@ -64,6 +90,18 @@ func copyPirschJs() {
 	}
 
 	if err := os.WriteFile("pirsch.js", content, 0755); err != nil {
+		panic(err)
+	}
+}
+
+func copyPirschEventsJs() {
+	content, err := os.ReadFile("../../js/pirsch-events.js")
+
+	if err != nil {
+		panic(err)
+	}
+
+	if err := os.WriteFile("pirsch-events.js", content, 0755); err != nil {
 		panic(err)
 	}
 }