+++ js/pirsch-events.js
@@ -14,7 +14,7 @@
 
     // ignore script on localhost
     const script = document.querySelector("#pirscheventsjs");
-    const dev = script.getAttribute("data-dev");
+    const dev = script.hasAttribute("data-dev");
 
     if (!dev && (/^localhost(.*)$|^127(\.[0-9]{1,3}){3}$/is.test(location.hostname) || location.protocol === "file:")) {
         console.warn("Pirsch ignores events on localhost. You can enable it by adding the data-dev attribute.");
@@ -91,18 +91,7 @@
             hostname = location.href.replace(location.hostname, hostname);
         }
 
-        const req = new XMLHttpRequest();
-        req.open("POST", endpoint);
-        req.setRequestHeader("Content-Type", "application/json;charset=UTF-8");
-        req.onload = () => {
-            if (req.status >= 200 && req.status < 300) {
-                resolve(req.response);
-            } else {
-                reject(req.statusText);
-            }
-        };
-        req.onerror = () => reject(req.statusText);
-        req.send(JSON.stringify({
+        if (navigator.sendBeacon(endpoint, JSON.stringify({
             client_id: clientID,
             url: hostname.substring(0, 1800),
             title: document.title,
@@ -112,6 +101,10 @@
             event_name: name,
             event_duration: options && options.duration && typeof options.duration === "number" ? options.duration : 0,
             event_meta: meta
-        }));
+        }))) {
+            resolve();
+        } else {
+            reject("error queuing event request");
+        }
     }
 })();